tf_bench/requirements-tf.txt
tensorboard==2.11.2
tensorflow==2.11.0
tf-models-official==2.11.3
tensorboard-plugin-profile==2.11.1
transformers==4.27.1
fire==0.5.0
